pyproject.toml
[tool.poetry]
name = "sdss-skymakercam"
version = "0.1.0-alpha.0"
description = "Skymaker camera with sdss-basecam"
authors = ["Florian Briegel <briegel@mpia.de>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/wasndas/skymakercam"
repository = "https://github.com/wasndas/skymakercam"
documentation = "https://sdss-skymakercam.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "skymakercam", from = "python" }
]
include = ["python/skymakercam/etc/*"]

[tool.poetry.build]
script = "build.py"
generate-setup-file = false

[tool.poetry.scripts]
singleframe = "skymakercam.__main__:main"
plotloop = "utils.plotloop:main"
proto = "proto.__main__:proto"

[tool.poetry.dependencies]
python = ">3.7,<3.10"
click-default-group = "^1.2.2"
daemonocle = "^1.1.1"
sdss-tree = ">=2.15.2"
sdss-access = ">=0.2.3"
sdsstools = ">=0.4.0"
matplotlib = ">=3.1.1"
astropy = ">=4.0"
astroquery = ">0.4"
healpy = ">=1.15"
sep = ">=1.2"
numpy = ">=1.0"
scipy = ">=1.7"
scikit_image = ">=0.18"
PyQt5 = ">=5.14"
sdss-clu = ">=1.2.0"
sdss-basecam = ">= 0.5"
keyboard = ">0.13"
asyncudp = ">=0.4"
transitions = ">0.8"
jsonpickle = "1.3"

[tool.poetry.dev-dependencies]
ipython = ">=7.11.0"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = "^3.3.1"
Sphinx = "^3.0.0"

[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "skymakercam"
known_sdss = ["sdsstools"]
balanced_wrapping = true
include_trailing_comma = false
lines_after_imports = 2
use_parentheses = true

[tool.pytest.ini_options]
addopts = "--cov skymakercam --cov-report xml --cov-report html --cov-report term"

[tool.coverage.run]
branch = true
include = ["python/skymakercam/*"]
omit = [
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
]

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=49.6.0"]
build-backend = "poetry.core.masonry.api"
